Deep-Learning-main/Project/ResumeIQ/requirements.txt
streamlit
langchain
langchain-groq
langchain-community
pypdf
python-docx
python-dotenv
